files updated for sqlhander
