+++ SQLHelper/SQLHandler/SQLAddUpdate.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace SQLHelper
+{
+    public class SQLAddUpdate
+    {
+        public SQLAddUpdate()
+        {
+        }
+    }
+}
