+++ SQLHelper/SQLHandler/SQLGet.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace SQLHelper
+{
+    public class SQLGet
+    {
+        public SQLGet()
+        {
+        }
+    }
+}
