+++ SQLHelper/SQLHandler/SQLGetList.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace SQLHelper
+{
+    public class SQLGetList
+    {
+        public SQLGetList()
+        {
+        }
+    }
+}
