+++ SQLHelper/SQLHandler/SQLRaw.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace SQLHelper
+{
+    public class SQLRaw
+    {
+        public SQLRaw()
+        {
+        }
+    }
+}
